+++ r-package/R/list_geobr.R
@@ -3,6 +3,9 @@
 #' @description
 #' Returns a data frame with all datasets available in the geobr package
 #'
+#' @param wide Whether the the output data frame should come in wide (`TRUE`) or
+#'        long format (`FALSE`).
+#'
 #' @return A `data.frame`
 #'
 #' @export
@@ -11,72 +14,157 @@
 #' @examplesIf identical(tolower(Sys.getenv("NOT_CRAN")), "true")
 #' df <- list_geobr()
 #'
-list_geobr <- function(){
+list_geobr <- function(wide = TRUE){
 
-# Get readme.md file
-tempf <- fs::path(fs::path_temp(), "readme.md")
+  # download metadata
+  metadata <- download_metadata2()
 
-# check if metadata has already been downloaded
-if (file.exists(tempf) & file.info(tempf)$size != 0) {
-  readme <- readLines(tempf, encoding = "UTF-8")
+  # select cols
+  tempdf <- metadata |>
+    dplyr::select(alias = geo, year) |>
+    unique()
 
-} else {
-  # download it and save to metadata
-  git_url = "https://raw.githubusercontent.com/ipeaGIT/geobr/master/README.md"
-
-  # test server connection
-  try(silent = TRUE,
-      check_con <- check_connection(git_url)
-      )
-  if(is.null(check_con) | isFALSE(check_con)){ return(invisible(NULL)) }
-
-  try( silent = TRUE,
-       downloaded_files <- curl::multi_download(
-         urls = git_url,
-         destfiles = tempf,
-         resume = TRUE,
-         progress = FALSE
-       )
-  )
-
-  readme <- readLines(tempf, encoding = "UTF-8")
-}
+  # reformat to wide
+  if (isTRUE(wide)) {
+    tempdf <- tempdf |>
+      dplyr::group_by(alias) |>
+      dplyr::summarise( year = paste0(year, collapse = ', '))
+  }
 
 
 
-# find start and end of table
-table_start <- grep("Available datasets:", readme) + 4
-table_end <- grep("Other functions", readme) -1
+  datasets <- structure(list(
+  Function = c(
+    "read_country",
+    "read_region",
+    "read_state",
+    "read_meso_region",
+    "read_micro_region",
+    "read_intermediate_region",
+    "read_immediate_region",
+    "read_municipality",
+    "read_municipal_seat",
+    "read_weighting_area",
+    "read_census_tract",
+    "read_statistical_grid",
+    "read_metro_area",
+    "read_urban_area",
+    "read_amazon",
+    "read_biomes",
+    "read_conservation_units",
+    "read_disaster_risk_area",
+    "read_indigenous_land",
+    "read_semiarid",
+    "read_health_facilities",
+    "read_health_region",
+    "read_neighborhood",
+    "read_schools",
+    "read_comparable_areas",
+    "read_urban_concentrations",
+    "read_pop_arrangements",
+    "read_favelas"
+    ),
+ geography = c(
+   "Country",
+   "Region",
+   "States",
+   "Meso region",
+   "Micro region",
+   "Intermediate region",
+   "Immediate region",
+   "Municipality",
+   "Municipality seats (sedes municipais)",
+   "Census weighting area (área de ponderação)",
+   "Census tract (setor censitário)",
+   "Statistical Grid (Gridded population)",
+   "Metropolitan areas",
+   "Urban footprints",
+   "Brazil's Legal Amazon",
+   "Biomes",
+   "Environmental Conservation Units",
+   "Disaster risk areas",
+   "Indigenous lands",
+   "Semi Arid region",
+   "Health facilities",
+   "Health regions and macro regions",
+   "Neighborhood limits",
+   "Schools",
+   "Historically comparable municipalities, aka Áreas mínimas comparáveis (AMCs)",
+   "Urban concentration areas (concentrações urbanas)",
+   "Population arrangements (arranjos populacionais)",
+   "Favelas and urban communities"
+ ),
+ source = c(
+   "IBGE",
+   "IBGE",
+   "IBGE",
+   "IBGE",
+   "IBGE",
+   "IBGE",
+   "IBGE",
+   "IBGE",
+   "IBGE",
+   "IBGE",
+   "IBGE",
+   "IBGE",
+   "IBGE",
+   "IBGE",
+   "MMA",
+   "IBGE",
+   "MMA",
+   "CEMADEN and IBGE",
+   "FUNAI","IBGE",
+   "CNES, DataSUS",
+   "DataSUS",
+   "IBGE",
+   "INEP",
+   "IBGE",
+   "IBGE",
+   "IBGE",
+   "IBGE"
+   ),
 
-# get table string in Markdown
-table_strig <- readme[table_start:table_end]
+ alias = c(
+   "country",
+   "regions",
+   "states",
+   "mesoregions",
+   "microregions",
+   "intermediateregions",
+   "immediateregions",
+   "municipalities",
+   "municipalseats",
+   "weightingareas",
+   "censustracts",
+   "statsgrid",
+   "metroareas",
+   "urbanareas",
+   "amazonialegal",
+   "biomes",
+   "conservationunits",
+   "disasterriskareas",
+   "indigenouslands",
+   "read_semiarid",
+   "healthfacilities",
+   "healthregions",
+   "neighborhoods",
+   "read_schools",
+   "read_comparable_areas",
+   "urbanconcentrations",
+   "poparrangements",
+   "favelas"
+ )
+ ),
+ row.names = c(NA, 28L),
+ class = "data.frame"
+ )
 
-# read table as a data.frame
-# suppressWarnings({ df <- readr::read_delim(I(table_strig),
-#                                            delim = '|',
-#                                            trim_ws = TRUE,
-#                                            show_col_types = FALSE,
-#                                            col_names = FALSE) })
-suppressWarnings({ df <- utils::read.table(text = I(table_strig),
-                                           header = FALSE,
-                                           sep = '|',
-                                           strip.white = TRUE,
-                                           colClasses = 'character',
-                                           row.names = NULL,
-                                           quote="")
-                                        })
 
-# clean colunms
-  df[[1]] <- NULL
-  df[[5]] <- NULL
 
-# remove 1st row with "-----"
-  df <- df[-1,]
-  df <- df[-1,]
-  rownames(df) <- 1:nrow(df)
 
-# rename columns
-colnames(df) <- c("function", "geography", "years", "source")
+  df <- dplyr::left_join(datasets, tempdf, by  = 'alias') |>
+    dplyr::arrange(alias) |>
+    dplyr::select(-alias)
 
-return(df)
+  return(df)
 }